Move branch and bound to a seperate module
